requirements.txt
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.26.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
easyocr>=1.7.1
pdfplumber>=0.11.0
google-cloud-vision>=3.7.0
pydantic>=2.5.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.12.0
jsonschema>=4.21.0
pytest>=8.0.0

